Sema: Split off Apple-specific parts of bidirectional_conversions.swift test
